feat: Complete the stretch tasks
In this commit, i extend the functionality of the Student Class by adding
a property called grade which is initialized to 0.
A graduate() is also added to the Student class. this method is used to check if a
student is ready to graduate or not by checking to see if the grade of a student
is greater than 70. if grade is greater than 70, isReadyToGraduate prop is set to
true and a statement is logged to the student.

LAMBDA CLASSES (STRETCH TASKS)